Rework the Settings page to support the secondary navigation
